Create Polymorphism
This program shows the use of the virtual declaration and class inheritance. Virtual declaration allows the object to be declared only when stated by the user. This alleviates the burden of having to load the whole program, and unused functions. Thus saving the amount of space the process uses. Class inheritance allows classes to share qualities of one another.